+++ core/ajax/config.ajax.php
@@ -19,13 +19,13 @@
 try {
 	require_once __DIR__ . '/../../core/php/core.inc.php';
 	include_file('core', 'authentification', 'php');
-	
+
 	if (!isConnect()) {
 		throw new Exception(__('401 - Accès non autorisé', __FILE__), -1234);
 	}
-	
-	ajax::init();
-	
+
+	ajax::init(array('uploadImage'));
+
 	if (init('action') == 'genApiKey') {
 		if (!isConnect('admin')) {
 			throw new Exception(__('401 - Accès non autorisé', __FILE__));
@@ -45,7 +45,7 @@ try {
 			ajax::success(config::byKey('api', init('plugin')));
 		}
 	}
-	
+
 	if (init('action') == 'getKey') {
 		$keys = init('key');
 		if ($keys == '') {
@@ -66,7 +66,7 @@ try {
 			ajax::success($return);
 		}
 	}
-	
+
 	if (init('action') == 'addKey') {
 		if (!isConnect('admin')) {
 			throw new Exception(__('401 - Accès non autorisé', __FILE__));
@@ -78,7 +78,7 @@ try {
 		}
 		ajax::success();
 	}
-	
+
 	if (init('action') == 'removeKey') {
 		unautorizedInDemo();
 		$keys = init('key');
@@ -96,7 +96,59 @@ try {
 		}
 		ajax::success();
 	}
-	
+
+	if (init('action') == 'uploadImage') {
+		if (!isConnect('admin')) {
+			throw new Exception(__('401 - Accès non autorisé', __FILE__));
+		}
+		unautorizedInDemo();
+
+		$page = init('id');
+		$key = 'interface::background::'.$page;
+		config::save($key, config::getDefaultConfiguration('core')['core'][$key], 'core');
+
+		if (!isset($_FILES['file'])) {
+			throw new Exception(__('Aucun fichier trouvé. Vérifiez le paramètre PHP (post size limit)', __FILE__));
+		}
+		$extension = strtolower(strrchr($_FILES['file']['name'], '.'));
+		if (!in_array($extension, array('.jpg', '.png'))) {
+			throw new Exception(__('Extension du fichier non valide (autorisé .jpg .png) : ', __FILE__) . $extension);
+		}
+		if (filesize($_FILES['file']['tmp_name']) > 5000000) {
+			throw new Exception(__('Le fichier est trop gros (maximum 5Mo)', __FILE__));
+		}
+
+		$uploaddir = realpath(__DIR__ . '/../../data/backgrounds');
+		if (!file_exists($uploaddir)) {
+			mkdir($uploaddir, 0777);
+		}
+
+
+		$filepath = $uploaddir.'/config_'.$page.$extension;
+		@unlink($filepath);
+		file_put_contents($filepath, file_get_contents($_FILES['file']['tmp_name']));
+		if (!file_exists($filepath)) {
+			throw new \Exception(__('Impossible de sauvegarder l\'image', __FILE__));
+		}
+
+		config::save($key, '/data/backgrounds/config_'.$page.$extension);
+		ajax::success();
+	}
+
+	if (init('action') == 'removeImage') {
+		if (!isConnect('admin')) {
+			throw new Exception(__('401 - Accès non autorisé', __FILE__));
+		}
+		unautorizedInDemo();
+
+		$page = init('id');
+		$key = 'interface::background::'.$page;
+		$filepath = config::byKey($key, 'core');
+		@unlink($filepath);
+		config::save($key, config::getDefaultConfiguration('core')['core'][$key], 'core');
+		ajax::success();
+	}
+
 	throw new Exception(__('Aucune méthode correspondante à : ', __FILE__) . init('action'));
 	/*     * *********Catch exeption*************** */
 } catch (Exception $e) {