+++ core/js/config.class.js
@@ -213,3 +213,22 @@ jeedom.config.remove = function (_params) {
     };
     $.ajax(paramsAJAX);
 };
+
+jeedom.config.removeImage = function (_params) {
+  var paramsRequired = ['id'];
+  var paramsSpecifics = {};
+  try {
+    jeedom.private.checkParamsRequired(_params || {}, paramsRequired);
+  } catch (e) {
+    (_params.error || paramsSpecifics.error || jeedom.private.default_params.error)(e);
+    return;
+  }
+  var params = $.extend({}, jeedom.private.default_params, paramsSpecifics, _params || {});
+  var paramsAJAX = jeedom.private.getParamsAJAX(params);
+  paramsAJAX.url = 'core/ajax/config.ajax.php';
+  paramsAJAX.data = {
+    action: 'removeImage',
+    id: _params.id
+  };
+  $.ajax(paramsAJAX);
+};